change name of CreateCohortButton
